Extract enums from types.ts
