+++ src/command_prechecks.ts
@@ -5,7 +5,6 @@ import {
 } from "./helpers/discord_utils";
 import GameSession from "./structures/game_session";
 import MessageContext from "./structures/message_context";
-import { GameType } from "./types";
 import { IPCLogger } from "./logger";
 import dbContext from "./database_context";
 import { state } from "./kmq_worker";
@@ -13,6 +12,7 @@ import MusicSession from "./structures/music_session";
 import { getTimeUntilRestart } from "./helpers/management_utils";
 import { isUserPremium } from "./helpers/game_utils";
 import PrecheckArgs from "./interfaces/precheck_args";
+import { GameType } from "./enums/game_type";
 
 const logger = new IPCLogger("command_prechecks");
