+++ src/commands/game_commands/hint.ts
@@ -10,7 +10,7 @@ import MessageContext from "../../structures/message_context";
 import { KmqImages } from "../../constants";
 import { getGuildPreference } from "../../helpers/game_utils";
 import { codeLine } from "../../helpers/utils";
-import { GuildTextableMessage, GameType } from "../../types";
+import { GuildTextableMessage } from "../../types";
 import GameSession from "../../structures/game_session";
 import EliminationScoreboard from "../../structures/elimination_scoreboard";
 import GuildPreference from "../../structures/guild_preference";
@@ -21,6 +21,7 @@ import Session from "../../structures/session";
 import HelpDocumentation from "../../interfaces/help";
 import CommandArgs from "../../interfaces/command_args";
 import { GuessModeType } from "../../enums/option_types/guess_mode_type";
+import { GameType } from "../../enums/game_type";
 
 const logger = new IPCLogger("hint");
