+++ src/commands/game_options/add.ts
@@ -9,7 +9,7 @@ import {
     getMatchingGroupNames,
 } from "../../helpers/game_utils";
 import { IPCLogger } from "../../logger";
-import { GameOption } from "../../types";
+import { GameOption } from "../../enums/game_option_name";
 import MessageContext from "../../structures/message_context";
 import { setIntersection } from "../../helpers/utils";
 import { GROUP_LIST_URL } from "../../constants";
