+++ src/commands/game_options/reset.ts
@@ -8,7 +8,7 @@ import {
 import MessageContext from "../../structures/message_context";
 import CommandPrechecks from "../../command_prechecks";
 import { GameOptionInternalToGameOption } from "../../structures/guild_preference";
-import { GameOption } from "../../types";
+import { GameOption } from "../../enums/game_option_name";
 import { state } from "../../kmq_worker";
 import HelpDocumentation from "../../interfaces/help";
 import CommandArgs from "../../interfaces/command_args";
