+++ src/commands/game_options/shuffle.ts
@@ -6,7 +6,7 @@ import {
     getDebugLogHeader,
     sendErrorMessage,
 } from "../../helpers/discord_utils";
-import { GameOption } from "../../types";
+import { GameOption } from "../../enums/game_option_name";
 import MessageContext from "../../structures/message_context";
 import CommandPrechecks from "../../command_prechecks";
 import { state } from "../../kmq_worker";
