+++ src/commands/game_options/special.ts
@@ -8,7 +8,7 @@ import { getGuildPreference, isUserPremium } from "../../helpers/game_utils";
 import { state } from "../../kmq_worker";
 import { IPCLogger } from "../../logger";
 import GuildPreference from "../../structures/guild_preference";
-import { GameOption } from "../../types";
+import { GameOption } from "../../enums/game_option_name";
 import MessageContext from "../../structures/message_context";
 import CommandPrechecks from "../../command_prechecks";
 import CommandArgs from "../../interfaces/command_args";
