+++ src/enums/env_type.ts
@@ -1,0 +1,7 @@
+export enum EnvType {
+    PROD = "production",
+    DEV = "development",
+    DRY_RUN = "dry-run",
+    CI = "ci",
+    TEST = "test",
+}
