+++ src/enums/game_option_name.ts
@@ -1,0 +1,23 @@
+export enum GameOption {
+    LIMIT = "Limit",
+    GROUPS = "Groups",
+    GENDER = "Gender",
+    CUTOFF = "Cutoff",
+    ARTIST_TYPE = "Artist Type",
+    ANSWER_TYPE = "Answer Type",
+    RELEASE_TYPE = "Release Type",
+    LANGUAGE_TYPE = "Language Type",
+    SUBUNIT_PREFERENCE = "Subunit Preference",
+    OST_PREFERENCE = "OST Preference",
+    MULTIGUESS = "Multiguess",
+    SHUFFLE_TYPE = "Shuffle",
+    SEEK_TYPE = "Seek Type",
+    SPECIAL_TYPE = "Special Type",
+    GUESS_MODE_TYPE = "Guess Mode",
+    GOAL = "Goal",
+    TIMER = "Timer",
+    DURATION = "Duration",
+    EXCLUDE = "Exclude",
+    INCLUDE = "Include",
+    FORCE_PLAY_SONG = "Force Play Song",
+}
