+++ src/enums/game_type.ts
@@ -1,0 +1,6 @@
+export enum GameType {
+    CLASSIC = "classic",
+    ELIMINATION = "elimination",
+    TEAMS = "teams",
+    COMPETITION = "competition",
+}
