+++ src/kmq_worker.ts
@@ -5,7 +5,6 @@ import path from "path";
 import { Campaign } from "patreon-discord";
 
 import { IPCLogger } from "./logger";
-import { EnvType } from "./types";
 import {
     registerClientEvents,
     registerIntervals,
@@ -22,6 +21,7 @@ import EvalCommand from "./commands/admin/eval";
 import LocalizationManager from "./helpers/localization_manager";
 import Session from "./structures/session";
 import { State } from "./interfaces/state";
+import { EnvType } from "./enums/env_type";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const logger = new IPCLogger("kmq");
