+++ src/seed/bootstrap.ts
@@ -6,7 +6,7 @@ import { updateKpopDatabase } from "./seed_db";
 import { IPCLogger } from "../logger";
 import { downloadAndConvertSongs } from "../scripts/download-new-songs";
 import { DatabaseContext, getNewConnection } from "../database_context";
-import { EnvType } from "../types";
+import { EnvType } from "../enums/env_type";
 
 const logger = new IPCLogger("bootstrap");
