+++ src/structures/guild_preference.ts
@@ -14,7 +14,6 @@ import dbContext from "../database_context";
 import { DEFAULT_ARTIST_TYPE } from "../commands/game_options/artisttype";
 import { DEFAULT_LANGUAGE } from "../commands/game_options/language";
 import { DEFAULT_SUBUNIT_PREFERENCE } from "../commands/game_options/subunits";
-import { GameOption } from "../types";
 import { DEFAULT_OST_PREFERENCE } from "../commands/game_options/ost";
 import { DEFAULT_RELEASE_TYPE } from "../commands/game_options/release";
 import { DEFAULT_MULTIGUESS_TYPE } from "../commands/game_options/multiguess";
@@ -34,6 +33,7 @@ import { GuessModeType } from "../enums/option_types/guess_mode_type";
 import { ArtistType } from "../enums/option_types/artist_type";
 import { Gender } from "../enums/option_types/gender";
 import { SeekType } from "../enums/option_types/seek_type";
+import { GameOption } from "../enums/game_option_name";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const logger = new IPCLogger("guild_preference");
